+++ src/components/Film/Film.tsx
@@ -1,7 +1,7 @@
 import styles from './Film.module.css';
 import clsx from 'clsx';
 import { useGetFilmQuery } from '../../store/Redux/api';
-import { FetchBaseQueryError } from '@reduxjs/toolkit/query';
+import type { FetchBaseQueryError } from '@reduxjs/toolkit/query';
 
 export default function Film({ film }: { film: string }) {
   const { data, isLoading, error } = useGetFilmQuery(film);
@@ -28,19 +28,19 @@ export default function Film({ film }: { film: string }) {
     <div className={styles.title}>
       <h2>{data?.title}</h2>
       <p>
-        <b>Director </b>:{data?.director}
+        <b>Director</b>: {data?.director}
       </p>
       <p>
-        <b>Producer </b>:{data?.producer}
+        <b>Producer</b>: {data?.producer}
       </p>
       <p>
-        <b>release </b>:{data?.release_date}
+        <b>release</b>: {data?.release_date}
       </p>
       <p>
-        <b>characters </b>:{data?.characters?.length}
+        <b>characters</b>: {data?.characters?.length}
       </p>
       <p>
-        <b>planets </b>:{data?.opening_crawl}
+        <b>planets</b>: {data?.opening_crawl}
       </p>
     </div>
   );
