+++ src/components/index.tsx
@@ -1,5 +1,6 @@
 export { default as ErrorBoundary } from './ErrorBoundary/ErrorBoundary.tsx';
 export { default as ErrorButton } from './ErrorButton/ErrorButton.tsx';
+export { default as ErrorPage } from './ErrorPage/ErrorPage.tsx';
 export { default as Search } from './Search/Search.tsx';
 export { default as CardList } from './CardList/CardList.tsx';
 export { default as Card } from './Card/Card.tsx';
