+++ src/types/types.tsx
@@ -29,11 +29,6 @@ export type Person = {
   results: Results[];
 };
 
-// export type State = {
-//   isLoading: boolean;
-//   pageLink: (page: string) => void;
-// } & Partial<Person>;
-
 export type Films = {
   title: string;
   opening_crawl: string;
